+++ apitanic/controller/user.py
@@ -1,4 +1,4 @@
-from apitanic.model.firebase import Firebase
+from apitanic.model.user import User
 from sanic.response import json, HTTPResponse
 from sanic.request import Request
 from sanic import Blueprint
@@ -7,7 +7,7 @@ from sanic_openapi import doc
 
 
 userBlueprint = Blueprint('user', url_prefix='user')
-firebase = Firebase()
+userModel = User()
 
 
 class UserController(HTTPMethodView):
@@ -18,11 +18,11 @@ class UserController(HTTPMethodView):
     @doc.produces({'data': {'results': list}})
     async def get(self, request: Request)-> HTTPResponse:
         token = request.headers.get('token')
-        user = firebase.get_user_id_by_token(token)
+        user = await userModel.get_user_id_by_token(token)
         if user is None:
             return json({'error': 'token not valid'})
         query_string = request.args['q'][0]
-        results = await firebase.search_user(query_string.lower())
+        results = await userModel.search_user(query_string.lower())
         return json({'data': {'results': results}})
 
     @doc.summary('Create a new account')
@@ -30,7 +30,7 @@ class UserController(HTTPMethodView):
     @doc.description('Create a user account and returns a token')
     @doc.produces({'data': {'token': str}})
     async def post(self, request: Request):
-        token = firebase.create_account(
+        token = userModel.create_account(
             request.json['email'],
             request.json['password'],
             request.json['displayName']
