+++ apitanic/model/__init__.py
@@ -1,0 +1,20 @@
+import os
+import firebase_admin
+from firebase_admin import credentials, firestore, auth
+from google.cloud.firestore_v1beta1 import CollectionReference
+
+cred = credentials.Certificate({
+    "type": os.getenv("type"),
+    "project_id": os.getenv("project_id"),
+    "private_key_id": os.getenv("private_key_id"),
+    "private_key": os.getenv("private_key").replace('\\n', '\n'),
+    "client_email": os.getenv("client_email"),
+    "client_id": os.getenv("client_id"),
+    "auth_uri": os.getenv("auth_uri"),
+    "token_uri": os.getenv("token_uri"),
+    "auth_provider_x509_cert_url": os.getenv("auth_provider_x509_cert_url"),
+    "client_x509_cert_url": os.getenv("client_x509_cert_url")
+})
+
+firebase_app = firebase_admin.initialize_app(cred)
+firebase_db = firestore.client()
