--- apitanic/model/friend.py
@@ -1,6 +1,0 @@
-from marshmallow import Schema, fields
-
-
-class FriendSchema(Schema):
-    friend_id = fields.Str()
-    user_id = fields.Str()
