+++ requirements.txt
@@ -3,26 +3,20 @@ cachetools==2.0.1
 certifi==2017.11.5
 chardet==3.0.4
 dill==0.2.7.1
-firebase-admin==2.7.0
+firebase-admin==2.8.0
 future==0.16.0
-gevent==1.2.2
-google-api-core==0.1.3
-google-auth==1.2.1
+google-api-core==0.1.4
+google-auth==1.3.0
 google-cloud-core==0.28.0
 google-cloud-firestore==0.28.0
 google-cloud-storage==1.6.0
 google-gax==0.15.16
 google-resumable-media==0.3.1
 googleapis-common-protos==1.5.3
-greenlet==0.4.12
-grpcio==1.8.2
-gunicorn==19.7.1
+grpcio==1.8.4
 httptools==0.0.10
 idna==2.6
-lockfile==0.12.2
 lxml==4.1.1
-marshmallow==2.15.0
-msgpack-python==0.4.8
 ply==3.8
 protobuf==3.5.1
 pyasn1==0.4.2
@@ -32,7 +26,6 @@ requests==2.18.4
 rsa==3.4.2
 sanic==0.7.0
 sanic-openapi==0.4.0
-Sanic-Plugins-Framework==0.5.0.dev20171225
 six==1.11.0
 ujson==1.35
 urllib3==1.22
